test(vket): 分割後の責務境界を整える
architecture reviewで指摘されたテスト配置と公開ページ命名を、振る舞いを保ったまま明確化するため。

検証: Vket全148件、AST一致、ruff 0.11.6、file-size-guard
